CYP-27 - Refactor Number One
Our first refactor :love_letter:
