+++ package.json
@@ -17,6 +17,7 @@
     "@testing-library/cypress": "^7.0.3",
     "cypress": "^6.3.0",
     "cypress-cucumber-preprocessor": "^4.0.0",
+    "eslint-config-prettier": "^7.2.0",
     "eslint-plugin-chai-friendly": "^0.6.0",
     "eslint-plugin-cypress": "^2.11.2",
     "husky": "^4.3.8",
